+++ ResourceFiles/GUI.cpp
@@ -134,7 +134,7 @@ gui::DropDownList::DropDownList(float x, float y, float width, float height, sf:
                                           sf::Color(70,70,70,200),sf::Color(150,150,150,200),sf::Color(20,20,20,200),
                                            sf::Color(255,255,255,200),sf::Color(255,255,255,255),sf::Color(20,20,20,50));
 
-    for (size_t i = 0; i < nrOfElements; i++)
+    for (unsigned i = 0; i < nrOfElements; i++)
     {
         this->list.push_back(new gui::Buttons(x,y + ((i+1) * height),width,height,
                                           &this->font,list[i], 12,
@@ -239,11 +239,11 @@ gui::TextureSelector::TextureSelector(float x, float y, float width, float heigh
 
     if (this->sheet.getGlobalBounds().width > this->bounds.getGlobalBounds().width)
     {
-        this->sheet.setTextureRect(sf::IntRect(0,0,this->bounds.getGlobalBounds().width, this->bounds.getGlobalBounds().height));
+        this->sheet.setTextureRect(sf::IntRect(0,0,static_cast<int>(this->bounds.getGlobalBounds().width), static_cast<int>(this->bounds.getGlobalBounds().height)));
     }
     else if (this->sheet.getGlobalBounds().height > this->bounds.getGlobalBounds().height)
     {
-        this->sheet.setTextureRect(sf::IntRect(0,0,this->bounds.getGlobalBounds().width, this->bounds.getGlobalBounds().height));
+        this->sheet.setTextureRect(sf::IntRect(0,0,static_cast<int>(this->bounds.getGlobalBounds().width), static_cast<int>(this->bounds.getGlobalBounds().height)));
     }
 
     this->selector.setPosition(x + 60,y);
